backend/requirements.txt
aerich==0.7.2
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dictdiffer==0.9.0
dnspython==2.6.1
easy_pyoc==0.2.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pypika-tortoise==0.1.6
pytest==8.3.3
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
rich==13.9.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
tomlkit==0.13.2
tortoise-orm==0.21.6
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
